Create a Base class for the methods used by multiple deploy tasks A few methods are used by both deploy_service and deploy_bouncer. Add a few comments over the code and update a few error messages.
