+++ lib/deploy_service.rb
@@ -1,4 +1,4 @@
-class DeployService
+class DeployService < DeployBase
   CONFIGS = YAML.load_file(File.join(__dir__, "..", "fastly.yaml"))
 
   def deploy!(argv)
@@ -21,7 +21,8 @@ class DeployService
       exit
     end
 
-    delete_ui_objects(service.id, version.number)
+    service_types = %w[healthcheck cache_settings request_settings response_object header gzip]
+    delete_ui_objects(service.id, version.number, service_types)
     upload_vcl(version, vcl)
     diff_vcl(service, version)
 
@@ -44,7 +45,7 @@ private
       nil
     end
 
-    raise "ERROR: Unknown configuration/environment combination: #{configuration} #{environment}. Check this combination exists in fastly.yaml." unless config_hash
+    raise "Error: Unknown configuration/environment combination: #{configuration} #{environment}. Check this combination exists in fastly.yaml." unless config_hash
 
     [configuration, environment, config_hash]
   end
@@ -56,28 +57,6 @@ private
     ref
   end
 
-  def get_dev_version(configuration)
-    # Sometimes the latest version isn't the development version.
-    version = configuration.version
-    version = version.clone if version.active?
-
-    version
-  end
-
-  def delete_ui_objects(service_id, version_number)
-    # Delete objects created by the UI. We want VCL to be the source of truth.
-    # Most of these don't have real objects in the Fastly API gem.
-    to_delete = %w[healthcheck cache_settings request_settings response_object header gzip]
-    to_delete.each do |type|
-      type_path = "/service/#{service_id}/version/#{version_number}/#{type}"
-      @fastly.client.get(type_path).map { |i| i["name"] }.each do |name|
-        puts "Deleting #{type}: #{name}"
-        resp = @fastly.client.delete("#{type_path}/#{ERB::Util.url_encode(name)}")
-        raise "ERROR: Failed to delete configuration" unless resp
-      end
-    end
-  end
-
   def modify_settings(version, ttl)
     settings = version.settings
     settings.settings.update(
@@ -87,41 +66,11 @@ private
     settings.save!
   end
 
-  def upload_vcl(version, contents)
-    vcl_name = "main"
-
-    begin
-      version.vcl(vcl_name) && version.delete_vcl(vcl_name)
-    rescue Fastly::Error => e
-      puts "Error: #{e.inspect}"
-    end
-
-    vcl = version.upload_vcl(vcl_name, contents)
-    @fastly.client.put("#{Fastly::VCL.put_path(vcl)}/main")
-  end
-
-  def diff_vcl(configuration, version_new)
-    version_current = configuration.versions.find(&:active?)
-
-    if version_current.nil?
-      raise "There are no active versions of this configuration"
-    end
-
-    diff = Diffy::Diff.new(
-      version_current.generated_vcl.content,
-      version_new.generated_vcl.content,
-      context: 3,
-    )
-
-    puts "Diff versions: #{version_current.number} -> #{version_new.number}"
-    puts diff.to_s(:color)
-  end
-
   def validate_config(version)
     is_valid_vcl, error_message = version.validate
 
     unless is_valid_vcl
-      raise "ERROR: Invalid configuration:\n #{error_message}"
+      raise "Error: Invalid configuration:\n #{error_message}"
     end
   end
 end
